refactor(app): refactor create ifarme
refactor create iframe && add reset style to iframe
