+++ example.html
@@ -94,7 +94,7 @@
 <h2>Basic Prebid.js Example</h2>
 
 <h5>Div-1</h5>
-<div id='div-1'>
+<div id='div-1' style="outline: 1px solid red">
     <script type='text/javascript'>
         googletag.cmd.push(function() {
             googletag.display('div-1');
