+++ src/debugHighlight.js
@@ -1,4 +1,4 @@
-import {createIframeProcess} from "./createIframe.js";
+import {createIframeProcess} from "./renderWinningBids.js";
 
 
 export const modifyFrame = (next, iframe)=>{
